@@ -3,7 +3,7 @@ class Sys::Admin::LdapGroupsController < Sys::Controller::Admin::Base
   
   def pre_dispatch
     return error_auth unless Core.user.has_auth?(:manager)
-    return render text: "LDAPサーバーに接続できません。", layout: true unless Core.ldap.connection
+    return render plain: 'LDAPサーバーに接続できません。', layout: true unless Core.ldap.connection
 
     if params[:parent] == '0'
       @parent  = nil
