@@ -532,7 +532,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
     if request.xhr?
       send_mdn_message(:manual)
-      return render text: ''
+      return render plain: ''
     else
       begin
         send_mdn_message(:manual)
@@ -579,7 +579,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
         redirect_to action: :show, id: params[:id]
       end
     else
-      render text: "OK"
+      render plain: 'OK'
     end
   end
 
@@ -634,7 +634,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
   def keep_params(options = {})
     if options[:mailbox].blank? && options[:controller].blank?
       keeps = params.slice(:page, :search, :s_keyword, :s_column, :s_status, :s_label,
-        :sort_key, :sort_order, :sort_starred, :new_window)
+        :sort_key, :sort_order, :sort_starred, :new_window).permit!
       options = options.reverse_merge(keeps)
     end
     options
@@ -648,7 +648,7 @@ class Webmail::Admin::MailsController < Webmail::Controller::Admin::Base
 
   def check_user_email
     if Core.current_user.email.blank?
-      return render text: "メールアドレスが登録されていません。", layout: true
+      return render plain: 'メールアドレスが登録されていません。', layout: true
     end
   end
 
