@@ -1,4 +1,4 @@
-class Webmail::Address < ActiveRecord::Base
+class Webmail::Address < ApplicationRecord
   include Sys::Model::Base
   include Sys::Model::Auth::Free
 
@@ -21,7 +21,6 @@ class Webmail::Address < ActiveRecord::Base
     before_validation :unquote_name
   end
 
-  scope :readable, ->(user = Core.user) { where(user_id: user.id) }
   scope :search, ->(params) {
     rel = all
     params.each do |k, vs|
